Partially address review notes.
